+++ Consts.cpp
@@ -6,7 +6,7 @@
 namespace Maths
 {
 
-auto Round(const double val, const int numDecimalPlaces) -> double
+auto Round(const float val, const int numDecimalPlaces) -> float
 {
   assert(numDecimalPlaces >= 0);
 
@@ -17,14 +17,14 @@ auto Round(const double val, const int numDecimalPlaces) -> double
   if (numDecimalPlaces != lastNumDecimalPlaces)
   {
     lastNumDecimalPlaces = numDecimalPlaces;
-    lastPowerOfTen       = std::pow(10.0F, numDecimalPlaces);
+    lastPowerOfTen       = std::pow(10.0, static_cast<double>(numDecimalPlaces));
     //      oneOnLastPowerOfTen= 1.0/lastPowerOfTen;
   }
 
-  const auto roundedVal = std::floor(val * lastPowerOfTen) / lastPowerOfTen;
+  const auto roundedVal = std::floor(static_cast<double>(val) * lastPowerOfTen) / lastPowerOfTen;
   //    if (std::fabs(roundedVal) < oneOnLastPowerOfTen) roundedVal= 0.0;
 
-  return roundedVal;
+  return static_cast<float>(roundedVal);
 }
 
 } // namespace Maths
